Fix import and casting bug

+++ src/idt/idt.c
@@ -1,6 +1,7 @@
 #include "idt.h"
 #include "config.h"
 #include "memory/memory.h"
+#include "kernel.h"
 
 struct idt_desc idt_descriptors[VIKASHOS_TOTAL_INTERRUPTS];
 struct idtr_desc idtr_descriptor;
@@ -23,7 +24,7 @@ void idt_set(int interrupt_no, void *address) {
 void idt_init() {
     memset(idt_descriptors, 0, sizeof(idt_descriptors));
     idtr_descriptor.limit = sizeof(idt_descriptors) - 1;
-    idtr_descriptor.base = idt_descriptors;
+    idtr_descriptor.base = (uint32_t)idt_descriptors;
 
     idt_set(0, idt_zero);
     idt_load(&idtr_descriptor);
